src/main/java/com/example/aesdemo/AesDemoApplication.java
package com.example.aesdemo;

import org.springframework.boot.CommandLineRunner;
import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;

import javax.crypto.*;
import javax.crypto.spec.IvParameterSpec;
import javax.crypto.spec.SecretKeySpec;
import java.security.InvalidAlgorithmParameterException;
import java.security.InvalidKeyException;
import java.security.NoSuchAlgorithmException;
import java.security.SecureRandom;
import java.util.Base64;

@SpringBootApplication
public class AesDemoApplication implements CommandLineRunner {

    public static void main(String[] args) {
        SpringApplication.run(AesDemoApplication.class, args);
    }

    @Override
    public void run(String... args) throws Exception {
        SecretKey secretKeyOrig = generateKey(256); // Generate a secret key
        String encryptionKey = Base64.getEncoder().encodeToString(secretKeyOrig.getEncoded()); // convert the secret key to string
        SecretKey secretKeyFromStr = convertStringToSecretKeyto(encryptionKey); // Convert the string to secret key
        IvParameterSpec ivParameterSpec = generateIv(); // Generate Initialization Vector
        String input = "This is a test text";
        String algorithm = "AES/CBC/PKCS5Padding";

        String cipherText = encrypt(algorithm, input, secretKeyFromStr, ivParameterSpec);
        String plainText = decrypt(algorithm, cipherText, secretKeyFromStr, ivParameterSpec);
        System.out.println(cipherText + "; " + plainText);
    }

    public static SecretKey convertStringToSecretKeyto(String encodedKey) {
        byte[] decodedKey = Base64.getDecoder().decode(encodedKey);
        SecretKey originalKey = new SecretKeySpec(decodedKey, 0, decodedKey.length, "AES");

        return originalKey;
    }

    public static SecretKey generateKey(int n) throws NoSuchAlgorithmException {
        KeyGenerator keyGenerator = KeyGenerator.getInstance("AES");
        keyGenerator.init(n);
        SecretKey key = keyGenerator.generateKey();

        return key;
    }

    public static IvParameterSpec generateIv() {
        byte[] iv = new byte[16];
        new SecureRandom().nextBytes(iv);

        return new IvParameterSpec(iv);
    }

    public static String encrypt(String algorithm, String input, SecretKey key, IvParameterSpec iv)
            throws NoSuchPaddingException, NoSuchAlgorithmException,
            InvalidAlgorithmParameterException, InvalidKeyException,
            BadPaddingException, IllegalBlockSizeException {
        Cipher cipher = Cipher.getInstance(algorithm);
        cipher.init(Cipher.ENCRYPT_MODE, key, iv);
        byte[] cipherText = cipher.doFinal(input.getBytes());

        return Base64.getEncoder().encodeToString(cipherText);
    }

    public static String decrypt(String algorithm, String cipherText, SecretKey key, IvParameterSpec iv)
            throws NoSuchPaddingException, NoSuchAlgorithmException,
            InvalidAlgorithmParameterException, InvalidKeyException,
            BadPaddingException, IllegalBlockSizeException {
        Cipher cipher = Cipher.getInstance(algorithm);
        cipher.init(Cipher.DECRYPT_MODE, key, iv);
        byte[] plainText = cipher.doFinal(Base64.getDecoder().decode(cipherText));

        return new String(plainText);
    }

}
